+++ Models/PeticionAgregar.py
@@ -69,9 +69,9 @@ class Peticion():
         subject = 'peticion para agregar'
         archivo = render_template("correo.html", id_platillo = id_platillo, subject = subject, platillo= platillo, descripcion= descripcion, precio_unitario =precio_unitario , tipo= tipo, image= link_image)
 
-        proveedor_correo = 'smtp.live.com: 587'
-        remitente = 'felipetabordasanchez@outlook.es'
-        password = 'qawsed123'
+        proveedor_correo = 'smtp.gmail.com: 587'
+        remitente = 'comapp.helloworld@gmail.com'
+        password = 'comapp123'
         #conexion a servidor
         servidor = smtplib.SMTP(proveedor_correo)
         servidor.starttls()
@@ -100,48 +100,53 @@ class Peticion():
         tipo = dataObject['tipo']
         img = dataObject['img']
         img_final = ''
-        
+
+        imagen_cambio = False
+        link_image = img
+
         if "data:image/gif;base64," in img:
             img_final = img[22::]
+            imagen_cambio = True
 
-        elif "data:image/jpeg;base64," in img:
+        if "data:image/jpeg;base64," in img:
             img_final = img[23::]
+            imagen_cambio = True
 
-        elif "data:image/png;base64," in img:
+        if "data:image/png;base64," in img:
             img_final = img[22::]
+            imagen_cambio = True
         
-        else:
-            return jsonify({}), 200
 
-        nameImg = str(platillo)
-        im = Image.open(BytesIO(base64.b64decode(img_final)))
-        im.save('{}'.format(nameImg+'.png'), 'PNG')
-        nombre_image = nameImg+'.png'
+        if imagen_cambio:
+            nameImg = str(platillo)
+            im = Image.open(BytesIO(base64.b64decode(img_final)))
+            im.save('{}'.format(nameImg+'.png'), 'PNG')
+            nombre_image = nameImg+'.png'
 
-        key = 'abcdefghijklmnopqrstuvwxyz0123456789ABCDEFGHIJKLMOPQRSTUVWXYZ+%$'
-        
-        string_random = ''.join(random.sample(key, 64))
+            key = 'abcdefghijklmnopqrstuvwxyz0123456789ABCDEFGHIJKLMOPQRSTUVWXYZ+%$'
+            
+            string_random = ''.join(random.sample(key, 64))
 
-        nombre = string_random + platillo + '.jpg'
+            nombre = string_random + platillo + '.jpg'
 
-        result = ''
+            result = ''
 
-        dbx = dropbox.Dropbox('i55bkV3doxoAAAAAAAAAAZHHYiUBwkXoHtTHt-S-1R7WmzjiR3CF1qH3LydQ4WEA')
+            dbx = dropbox.Dropbox('i55bkV3doxoAAAAAAAAAAZHHYiUBwkXoHtTHt-S-1R7WmzjiR3CF1qH3LydQ4WEA')
 
-        with open(nombre_image, 'rb') as f:
-            result = dbx.files_upload(f.read(), '/ComApp/Menu/' + nombre)
+            with open(nombre_image, 'rb') as f:
+                result = dbx.files_upload(f.read(), '/ComApp/Menu/' + nombre)
 
-        os.remove(nombre_image)
+            os.remove(nombre_image)
 
-        link = dbx.sharing_create_shared_link(path='/ComApp/Menu/' + nombre)
+            link = dbx.sharing_create_shared_link(path='/ComApp/Menu/' + nombre)
 
-        link_image = link.url.replace('?dl=0', '?dl=1')
+            link_image = link.url.replace('?dl=0', '?dl=1')
 
         subject = 'peticion para editar'
         archivo = render_template("correo.html", id_platillo = id_platillo, subject = subject, platillo= platillo, descripcion= descripcion, precio_unitario =precio_unitario , tipo= tipo, image= link_image  )
-        proveedor_correo = 'smtp.live.com: 587'
-        remitente = 'felipetabordasanchez@outlook.es'
-        password = 'qawsed123'
+        proveedor_correo = 'smtp.gmail.com: 587'
+        remitente = 'comapp.helloworld@gmail.com'
+        password = 'comapp123'
         #conexion a servidor
         servidor = smtplib.SMTP(proveedor_correo)
         servidor.starttls()
@@ -171,9 +176,9 @@ class Peticion():
         subject = 'peticion para Eliminar'
         archivo = render_template("correo.html", id_platillo = id_platillo, subject = subject, platillo= platillo, descripcion= descripcion, precio_unitario =precio_unitario , tipo= tipo)
 
-        proveedor_correo = 'smtp.live.com: 587'
-        remitente = 'felipetabordasanchez@outlook.es'
-        password = 'qawsed123'
+        proveedor_correo = 'smtp.gmail.com: 587'
+        remitente = 'comapp.helloworld@gmail.com'
+        password = 'comapp123'
         #conexion a servidor
         servidor = smtplib.SMTP(proveedor_correo)
         servidor.starttls()
